refactor change component TooltipButton -> TooltipToggleButton
- TooltipToggleButton move to folder src/components/editor -> src/components/common
- Remove custom interface and import ToogleButtonProps in MUI
